# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Admin Backend API",
        "description": "RESTful API for managing portfolio admin dashboard content with Clean Architecture implementation",
        "contact": {
            "name": "Nhật Anh Dev",
            "email": "admin@nhatanh.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "description": "Retrieve paginated list of projects with filtering options",
                "operationId": "c1725e676a2c746e33b7655c04753e65",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by project category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "web"
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "description": "Filter by featured status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in project titles and descriptions",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Laravel"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projects retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title_vi": {
                                                                "type": "string",
                                                                "example": "Ứng dụng E-commerce"
                                                            },
                                                            "title_en": {
                                                                "type": "string",
                                                                "example": "E-commerce Application"
                                                            },
                                                            "description_vi": {
                                                                "type": "string",
                                                                "example": "Ứng dụng thương mại điện tử hiện đại"
                                                            },
                                                            "description_en": {
                                                                "type": "string",
                                                                "example": "Modern e-commerce application"
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "example": "https://example.com/project1.jpg"
                                                            },
                                                            "link": {
                                                                "type": "string",
                                                                "example": "https://project1.com"
                                                            },
                                                            "technologies": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "Laravel"
                                                                }
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "web"
                                                            },
                                                            "featured": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "order": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}